+++ lib/pages/myPage/mypage_edit.dart
@@ -4,7 +4,7 @@ import 'package:blurting/Utils/provider.dart';
 import 'package:blurting/main.dart';
 import 'package:blurting/pages/mypage/utils.dart';
 import 'package:blurting/utils/util_widget.dart';
-import 'package:blurting/signup_questions/active_place_search.dart';
+import 'package:blurting/pages/signup_questions/active_place_search.dart';
 import 'package:blurting/mainapp.dart';
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
