+++ lib/pages/policy/policy_one.dart
@@ -1,17 +1,9 @@
-import 'package:blurting/mainapp.dart';
-import 'package:blurting/pages/useguide/useguide_done.dart';
-import 'package:blurting/pages/policy/policyFive.dart';
-import 'package:blurting/pages/policy/policyFour.dart';
-import 'package:blurting/pages/policy/policyThree.dart';
-import 'package:blurting/pages/policy/policyTwo.dart';
-import 'package:blurting/pages/useguide/useguidepageone.dart';
-import 'package:blurting/signup_questions/phone_number.dart';
-import 'package:blurting/token.dart';
+import 'package:blurting/pages/policy/policy_five.dart';
+import 'package:blurting/pages/policy/policy_four.dart';
+import 'package:blurting/pages/policy/policy_three.dart';
+import 'package:blurting/pages/policy/policy_two.dart';
+import 'package:blurting/pages/useguide/useguide_page_one.dart';
 import 'package:flutter/material.dart';
-import 'package:blurting/config/app_config.dart';
-import 'package:http/http.dart' as http;
-
-import 'package:blurting/utils/util_widget.dart';
 
 void main() {
   runApp(MyApp());
