+++ lib/pages/signup_questions/active_place.dart
@@ -4,10 +4,10 @@ import 'package:blurting/utils/util_widget.dart';
 import 'package:blurting/Utils/provider.dart';
 import 'package:http/http.dart' as http;
 import 'package:location/location.dart';
-import 'package:blurting/signup_questions/Utils.dart';
+import 'package:blurting/pages/signup_questions/Utils.dart';
 import 'package:blurting/token.dart';
-import 'package:blurting/signup_questions/religion.dart';
-import '../config/app_config.dart';
+import 'package:blurting/pages/signup_questions/religion.dart';
+import 'package:blurting/config/app_config.dart';
 import 'active_place_search.dart';
 
 class ActivePlacePage extends StatefulWidget {
