+++ lib/pages/signup_questions/height.dart
@@ -3,9 +3,9 @@ import 'package:blurting/config/app_config.dart';
 import 'package:http/http.dart' as http;
 import 'package:flutter/material.dart';
 import 'package:blurting/Utils/provider.dart';
-import 'package:blurting/signup_questions/Utils.dart';
+import 'package:blurting/pages/signup_questions/Utils.dart';
 import 'package:blurting/token.dart'; // token.dart를 임포트
-import 'package:blurting/signup_questions/major.dart'; // major.dart를 임포트
+import 'package:blurting/pages/signup_questions/major.dart'; // major.dart를 임포트
 import 'package:blurting/utils/util_widget.dart';
 
 final labels = ['안 핌', '가끔', '자주', '매일'];
@@ -143,7 +143,8 @@ class _HeightPageState extends State<HeightPage>
                 SizedBox(
                   height: 25,
                 ),
-                Center(child: ProgressBar(context, _progressAnimation!, _gender!)),
+                Center(
+                    child: ProgressBar(context, _progressAnimation!, _gender!)),
                 SizedBox(
                   height: 50,
                 ),
