+++ lib/pages/signup_questions/hobby.dart
@@ -1,58 +1,25 @@
 import 'dart:convert';
+import 'package:blurting/Utils/provider.dart';
 import 'package:flutter/material.dart';
 import 'package:blurting/token.dart';
-import 'package:blurting/signup_questions/utils.dart';
-import 'package:blurting/signup_questions/hobby.dart';
-import 'package:blurting/utils/util_widget.dart';
-import 'package:blurting/Utils/provider.dart';
+import 'package:blurting/pages/signup_questions/Utils.dart';
 import 'package:blurting/config/app_config.dart';
+import 'image.dart';
 import 'package:http/http.dart' as http;
+import 'package:blurting/utils/util_widget.dart';
 
-class PersonalityPage extends StatefulWidget {
+class HobbyPage extends StatefulWidget {
   final String selectedGender;
 
-  PersonalityPage({super.key, required this.selectedGender});
+  HobbyPage({required this.selectedGender});
   @override
-  _PersonalityPageState createState() => _PersonalityPageState();
+  HobbyPageState createState() => HobbyPageState();
 }
 
-class _PersonalityPageState extends State<PersonalityPage>
+class HobbyPageState extends State<HobbyPage>
     with SingleTickerProviderStateMixin {
   AnimationController? _animationController;
   Animation<double>? _progressAnimation;
-
-  Future<void> _increaseProgressAndNavigate() async {
-    await _animationController!.forward();
-    Navigator.of(context).push(
-      PageRouteBuilder(
-        pageBuilder: (context, animation, secondaryAnimation) =>
-            HobbyPage(selectedGender: widget.selectedGender),
-        transitionsBuilder: (context, animation, secondaryAnimation, child) {
-          return FadeTransition(opacity: animation, child: child);
-        },
-      ),
-    );
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    _animationController = AnimationController(
-      duration: Duration(milliseconds: 400), // 애니메이션의 지속 시간 설정
-      vsync: this,
-    );
-
-    _progressAnimation = Tween<double>(
-      begin: 10 / 15, // 시작 너비 (30%)
-      end: 11 / 15, // 종료 너비 (40%)
-    ).animate(
-        CurvedAnimation(parent: _animationController!, curve: Curves.easeInOut))
-      ..addListener(() {
-        setState(() {});
-      });
-  }
-
-  bool IsValid = false;
 
   List<bool> isValidList = [
     false,
@@ -71,26 +38,24 @@ class _PersonalityPageState extends State<PersonalityPage>
     false,
     false
   ];
-
   List<String> selectedCharacteristics = [];
 
   List<String> characteristic = [
-    "개성적인",
-    "유교중시",
-    "열정적인",
-    "귀여운",
-    "상냥한",
-    "감성적인",
-    "낙천적인",
-    "유머있는",
-    "차분한",
-    "집돌이",
-    "섬세한",
-    "오타쿠",
-    "MZ",
-    "갓생러"
+    "애니",
+    "그림그리기",
+    "술",
+    "영화/드라마",
+    "여행",
+    "요리",
+    "자기계발",
+    "독서",
+    "게임",
+    "노래듣기",
+    "봉사활동",
+    "운동",
+    "노래부르기",
+    "산책"
   ];
-
   void updateSelectedCharacteristics() {
     // 임시 리스트를 생성하여 선택된 특성들을 저장
     List<String> tempSelectedCharacteristics = [];
@@ -108,9 +73,9 @@ class _PersonalityPageState extends State<PersonalityPage>
     });
   }
 
-  Widget customPersonalityCheckBox(String hobbyText, int index, width, height) {
+  Widget customHobbyCheckbox(String hobbyText, int index, width, height) {
     return Container(
-      width: width * 0.42,
+      width: width * 0.44,
       height: height * 0.048,
       child: Row(
         mainAxisSize: MainAxisSize.min,
@@ -144,7 +109,7 @@ class _PersonalityPageState extends State<PersonalityPage>
                 color: mainColor.black,
                 fontFamily: 'Pretendard',
                 fontWeight: FontWeight.w500,
-                fontSize: 20,
+                fontSize: 18.6,
               ),
             ),
           ),
@@ -153,6 +118,20 @@ class _PersonalityPageState extends State<PersonalityPage>
     );
   }
 
+  Future<void> _increaseProgressAndNavigate() async {
+    await _animationController!.forward();
+    Navigator.of(context).push(
+      PageRouteBuilder(
+        pageBuilder: (context, animation, secondaryAnimation) =>
+            ImagePage(selectedGender: widget.selectedGender),
+        transitionsBuilder: (context, animation, secondaryAnimation, child) {
+          return FadeTransition(opacity: animation, child: child);
+        },
+      ),
+    );
+  }
+
+  bool IsValid = false;
   @override
   void IsSelected(int index) {
     var true_length = isValidList.where((item) => item == true).length;
@@ -169,23 +148,42 @@ class _PersonalityPageState extends State<PersonalityPage>
     }
   }
 
+  @override
+  void initState() {
+    super.initState();
+
+    _animationController = AnimationController(
+      duration: Duration(milliseconds: 600), // 애니메이션의 지속 시간 설정
+      vsync: this,
+    );
+
+    _progressAnimation = Tween<double>(
+      begin: 11 / 15, // 시작 너비 (30%)
+      end: 12 / 15, // 종료 너비 (40%)
+    ).animate(
+        CurvedAnimation(parent: _animationController!, curve: Curves.easeInOut))
+      ..addListener(() {
+        setState(() {});
+      });
+  }
+
   Future<void> _sendPostRequest() async {
     var url = Uri.parse(API.signup);
-    updateSelectedCharacteristics();
 
-    if (selectedCharacteristics.length > 4) {
-      showSnackBar(context, "성격 선택은 4개까지 가능합니다.");
-      return;
-    }
     String savedToken = await getToken();
 
+    updateSelectedCharacteristics();
+    if (selectedCharacteristics.length > 4) {
+      showSnackBar(context, "취미 선택은 4개까지 가능합니다.");
+      return;
+    }
     var response = await http.post(
       url,
       headers: <String, String>{
         'Content-Type': 'application/json; charset=UTF-8',
         'Authorization': 'Bearer $savedToken',
       },
-      body: json.encode({"character": selectedCharacteristics}), // JSON 형태로 인코딩
+      body: json.encode({"hobby": selectedCharacteristics}), // JSON 형태로 인코딩
     );
 
     if (response.statusCode == 200 || response.statusCode == 201) {
@@ -230,7 +228,7 @@ class _PersonalityPageState extends State<PersonalityPage>
         body: Padding(
           padding: const EdgeInsets.all(16.0),
           child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
+            crossAxisAlignment: CrossAxisAlignment.stretch,
             children: <Widget>[
               SizedBox(
                 height: 25,
@@ -240,23 +238,21 @@ class _PersonalityPageState extends State<PersonalityPage>
               SizedBox(
                 height: 50,
               ),
-              smallTitleQuestion("당신의 성격은 어떠신가요?"),
+              smallTitleQuestion("당신의 취미는 무엇인가요?"),
               SizedBox(height: 30),
               Row(
                 mainAxisAlignment: MainAxisAlignment.center, // 가로축 중앙 정렬
                 children: [
-                  customPersonalityCheckBox('개성있는', 0, width, height),
-                  customPersonalityCheckBox('유교중시', 1, width, height),
+                  customHobbyCheckbox('🍢애니', 0, width, height),
+                  customHobbyCheckbox('🎨그림그리기', 1, width, height),
                 ],
               ),
-              SizedBox(
-                height: 10,
-              ),
+              SizedBox(height: 10),
               Row(
                 mainAxisAlignment: MainAxisAlignment.center, // 가로축 중앙 정렬
                 children: [
-                  customPersonalityCheckBox('열정적인', 2, width, height),
-                  customPersonalityCheckBox('귀여운', 3, width, height),
+                  customHobbyCheckbox('🍻술', 2, width, height),
+                  customHobbyCheckbox('🎞️영화/드라마', 3, width, height),
                 ],
               ),
               SizedBox(
@@ -265,8 +261,8 @@ class _PersonalityPageState extends State<PersonalityPage>
               Row(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: [
-                  customPersonalityCheckBox('상냥한', 4, width, height),
-                  customPersonalityCheckBox('감성적인', 5, width, height),
+                  customHobbyCheckbox('✈️여행', 4, width, height),
+                  customHobbyCheckbox('🧑‍🍳요리', 5, width, height),
                 ],
               ),
               SizedBox(
@@ -275,8 +271,8 @@ class _PersonalityPageState extends State<PersonalityPage>
               Row(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: [
-                  customPersonalityCheckBox('낙천적인', 6, width, height),
-                  customPersonalityCheckBox('유머있는', 7, width, height),
+                  customHobbyCheckbox('🤓자기계발', 6, width, height),
+                  customHobbyCheckbox('📚독서', 7, width, height),
                 ],
               ),
               SizedBox(
@@ -285,8 +281,8 @@ class _PersonalityPageState extends State<PersonalityPage>
               Row(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: [
-                  customPersonalityCheckBox('차분한', 8, width, height),
-                  customPersonalityCheckBox('집돌이', 9, width, height),
+                  customHobbyCheckbox('🎮게임', 8, width, height),
+                  customHobbyCheckbox('🎧노래듣기', 9, width, height),
                 ],
               ),
               SizedBox(
@@ -295,8 +291,8 @@ class _PersonalityPageState extends State<PersonalityPage>
               Row(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: [
-                  customPersonalityCheckBox('섬세한', 10, width, height),
-                  customPersonalityCheckBox('오타쿠', 11, width, height),
+                  customHobbyCheckbox('🕊️봉사활동', 10, width, height),
+                  customHobbyCheckbox('🏃운동', 11, width, height),
                 ],
               ),
               SizedBox(
@@ -305,11 +301,11 @@ class _PersonalityPageState extends State<PersonalityPage>
               Row(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: [
-                  customPersonalityCheckBox('MZ', 12, width, height),
-                  customPersonalityCheckBox('갓생러', 13, width, height),
+                  customHobbyCheckbox('🎤노래부르기', 12, width, height),
+                  customHobbyCheckbox('🚶‍산책', 13, width, height),
                 ],
               ),
-              SizedBox(height: 26),
+              SizedBox(height: 10),
               Container(
                 width: 180,
                 height: 12,
@@ -323,7 +319,7 @@ class _PersonalityPageState extends State<PersonalityPage>
                     ),
                     children: [
                       TextSpan(
-                        text: '*성격을 최대 ',
+                        text: '*취미는 최대 ',
                       ),
                       TextSpan(
                         text: '4개',
