+++ lib/pages/signup_questions/major_list.dart
