+++ lib/pages/signup_questions/personality.dart
@@ -1,25 +1,58 @@
 import 'dart:convert';
-import 'package:blurting/Utils/provider.dart';
 import 'package:flutter/material.dart';
 import 'package:blurting/token.dart';
-import 'package:blurting/signup_questions/Utils.dart';
-import '../config/app_config.dart';
-import 'image.dart';
-import 'package:http/http.dart' as http;
+import 'package:blurting/pages/signup_questions/utils.dart';
+import 'package:blurting/pages/signup_questions/hobby.dart';
 import 'package:blurting/utils/util_widget.dart';
+import 'package:blurting/Utils/provider.dart';
+import 'package:blurting/config/app_config.dart';
+import 'package:http/http.dart' as http;
 
-class HobbyPage extends StatefulWidget {
+class PersonalityPage extends StatefulWidget {
   final String selectedGender;
 
-  HobbyPage({required this.selectedGender});
+  PersonalityPage({super.key, required this.selectedGender});
   @override
-  HobbyPageState createState() => HobbyPageState();
+  _PersonalityPageState createState() => _PersonalityPageState();
 }
 
-class HobbyPageState extends State<HobbyPage>
+class _PersonalityPageState extends State<PersonalityPage>
     with SingleTickerProviderStateMixin {
   AnimationController? _animationController;
   Animation<double>? _progressAnimation;
+
+  Future<void> _increaseProgressAndNavigate() async {
+    await _animationController!.forward();
+    Navigator.of(context).push(
+      PageRouteBuilder(
+        pageBuilder: (context, animation, secondaryAnimation) =>
+            HobbyPage(selectedGender: widget.selectedGender),
+        transitionsBuilder: (context, animation, secondaryAnimation, child) {
+          return FadeTransition(opacity: animation, child: child);
+        },
+      ),
+    );
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    _animationController = AnimationController(
+      duration: Duration(milliseconds: 400), // 애니메이션의 지속 시간 설정
+      vsync: this,
+    );
+
+    _progressAnimation = Tween<double>(
+      begin: 10 / 15, // 시작 너비 (30%)
+      end: 11 / 15, // 종료 너비 (40%)
+    ).animate(
+        CurvedAnimation(parent: _animationController!, curve: Curves.easeInOut))
+      ..addListener(() {
+        setState(() {});
+      });
+  }
+
+  bool IsValid = false;
 
   List<bool> isValidList = [
     false,
@@ -38,24 +71,26 @@ class HobbyPageState extends State<HobbyPage>
     false,
     false
   ];
+
   List<String> selectedCharacteristics = [];
 
   List<String> characteristic = [
-    "애니",
-    "그림그리기",
-    "술",
-    "영화/드라마",
-    "여행",
-    "요리",
-    "자기계발",
-    "독서",
-    "게임",
-    "노래듣기",
-    "봉사활동",
-    "운동",
-    "노래부르기",
-    "산책"
+    "개성적인",
+    "유교중시",
+    "열정적인",
+    "귀여운",
+    "상냥한",
+    "감성적인",
+    "낙천적인",
+    "유머있는",
+    "차분한",
+    "집돌이",
+    "섬세한",
+    "오타쿠",
+    "MZ",
+    "갓생러"
   ];
+
   void updateSelectedCharacteristics() {
     // 임시 리스트를 생성하여 선택된 특성들을 저장
     List<String> tempSelectedCharacteristics = [];
@@ -73,9 +108,9 @@ class HobbyPageState extends State<HobbyPage>
     });
   }
 
-  Widget customHobbyCheckbox(String hobbyText, int index, width, height) {
+  Widget customPersonalityCheckBox(String hobbyText, int index, width, height) {
     return Container(
-      width: width * 0.44,
+      width: width * 0.42,
       height: height * 0.048,
       child: Row(
         mainAxisSize: MainAxisSize.min,
@@ -109,7 +144,7 @@ class HobbyPageState extends State<HobbyPage>
                 color: mainColor.black,
                 fontFamily: 'Pretendard',
                 fontWeight: FontWeight.w500,
-                fontSize: 18.6,
+                fontSize: 20,
               ),
             ),
           ),
@@ -118,20 +153,6 @@ class HobbyPageState extends State<HobbyPage>
     );
   }
 
-  Future<void> _increaseProgressAndNavigate() async {
-    await _animationController!.forward();
-    Navigator.of(context).push(
-      PageRouteBuilder(
-        pageBuilder: (context, animation, secondaryAnimation) =>
-            ImagePage(selectedGender: widget.selectedGender),
-        transitionsBuilder: (context, animation, secondaryAnimation, child) {
-          return FadeTransition(opacity: animation, child: child);
-        },
-      ),
-    );
-  }
-
-  bool IsValid = false;
   @override
   void IsSelected(int index) {
     var true_length = isValidList.where((item) => item == true).length;
@@ -148,42 +169,23 @@ class HobbyPageState extends State<HobbyPage>
     }
   }
 
-  @override
-  void initState() {
-    super.initState();
-
-    _animationController = AnimationController(
-      duration: Duration(milliseconds: 600), // 애니메이션의 지속 시간 설정
-      vsync: this,
-    );
-
-    _progressAnimation = Tween<double>(
-      begin: 11 / 15, // 시작 너비 (30%)
-      end: 12 / 15, // 종료 너비 (40%)
-    ).animate(
-        CurvedAnimation(parent: _animationController!, curve: Curves.easeInOut))
-      ..addListener(() {
-        setState(() {});
-      });
-  }
-
   Future<void> _sendPostRequest() async {
     var url = Uri.parse(API.signup);
-
-    String savedToken = await getToken();
-
     updateSelectedCharacteristics();
+
     if (selectedCharacteristics.length > 4) {
-      showSnackBar(context, "취미 선택은 4개까지 가능합니다.");
+      showSnackBar(context, "성격 선택은 4개까지 가능합니다.");
       return;
     }
+    String savedToken = await getToken();
+
     var response = await http.post(
       url,
       headers: <String, String>{
         'Content-Type': 'application/json; charset=UTF-8',
         'Authorization': 'Bearer $savedToken',
       },
-      body: json.encode({"hobby": selectedCharacteristics}), // JSON 형태로 인코딩
+      body: json.encode({"character": selectedCharacteristics}), // JSON 형태로 인코딩
     );
 
     if (response.statusCode == 200 || response.statusCode == 201) {
@@ -228,30 +230,33 @@ class HobbyPageState extends State<HobbyPage>
         body: Padding(
           padding: const EdgeInsets.all(16.0),
           child: Column(
-            crossAxisAlignment: CrossAxisAlignment.stretch,
+            crossAxisAlignment: CrossAxisAlignment.start,
             children: <Widget>[
               SizedBox(
                 height: 25,
               ),
-              Center(child: ProgressBar(context, _progressAnimation!, _gender!)),
+              Center(
+                  child: ProgressBar(context, _progressAnimation!, _gender!)),
               SizedBox(
                 height: 50,
               ),
-              smallTitleQuestion("당신의 취미는 무엇인가요?"),
+              smallTitleQuestion("당신의 성격은 어떠신가요?"),
               SizedBox(height: 30),
               Row(
                 mainAxisAlignment: MainAxisAlignment.center, // 가로축 중앙 정렬
                 children: [
-                  customHobbyCheckbox('🍢애니', 0, width, height),
-                  customHobbyCheckbox('🎨그림그리기', 1, width, height),
+                  customPersonalityCheckBox('개성있는', 0, width, height),
+                  customPersonalityCheckBox('유교중시', 1, width, height),
                 ],
               ),
-              SizedBox(height: 10),
+              SizedBox(
+                height: 10,
+              ),
               Row(
                 mainAxisAlignment: MainAxisAlignment.center, // 가로축 중앙 정렬
                 children: [
-                  customHobbyCheckbox('🍻술', 2, width, height),
-                  customHobbyCheckbox('🎞️영화/드라마', 3, width, height),
+                  customPersonalityCheckBox('열정적인', 2, width, height),
+                  customPersonalityCheckBox('귀여운', 3, width, height),
                 ],
               ),
               SizedBox(
@@ -260,8 +265,8 @@ class HobbyPageState extends State<HobbyPage>
               Row(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: [
-                  customHobbyCheckbox('✈️여행', 4, width, height),
-                  customHobbyCheckbox('🧑‍🍳요리', 5, width, height),
+                  customPersonalityCheckBox('상냥한', 4, width, height),
+                  customPersonalityCheckBox('감성적인', 5, width, height),
                 ],
               ),
               SizedBox(
@@ -270,8 +275,8 @@ class HobbyPageState extends State<HobbyPage>
               Row(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: [
-                  customHobbyCheckbox('🤓자기계발', 6, width, height),
-                  customHobbyCheckbox('📚독서', 7, width, height),
+                  customPersonalityCheckBox('낙천적인', 6, width, height),
+                  customPersonalityCheckBox('유머있는', 7, width, height),
                 ],
               ),
               SizedBox(
@@ -280,8 +285,8 @@ class HobbyPageState extends State<HobbyPage>
               Row(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: [
-                  customHobbyCheckbox('🎮게임', 8, width, height),
-                  customHobbyCheckbox('🎧노래듣기', 9, width, height),
+                  customPersonalityCheckBox('차분한', 8, width, height),
+                  customPersonalityCheckBox('집돌이', 9, width, height),
                 ],
               ),
               SizedBox(
@@ -290,8 +295,8 @@ class HobbyPageState extends State<HobbyPage>
               Row(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: [
-                  customHobbyCheckbox('🕊️봉사활동', 10, width, height),
-                  customHobbyCheckbox('🏃운동', 11, width, height),
+                  customPersonalityCheckBox('섬세한', 10, width, height),
+                  customPersonalityCheckBox('오타쿠', 11, width, height),
                 ],
               ),
               SizedBox(
@@ -300,11 +305,11 @@ class HobbyPageState extends State<HobbyPage>
               Row(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: [
-                  customHobbyCheckbox('🎤노래부르기', 12, width, height),
-                  customHobbyCheckbox('🚶‍산책', 13, width, height),
+                  customPersonalityCheckBox('MZ', 12, width, height),
+                  customPersonalityCheckBox('갓생러', 13, width, height),
                 ],
               ),
-              SizedBox(height: 10),
+              SizedBox(height: 26),
               Container(
                 width: 180,
                 height: 12,
@@ -318,7 +323,7 @@ class HobbyPageState extends State<HobbyPage>
                     ),
                     children: [
                       TextSpan(
-                        text: '*취미는 최대 ',
+                        text: '*성격을 최대 ',
                       ),
                       TextSpan(
                         text: '4개',
