+++ lib/pages/signup_questions/religion.dart
@@ -1,8 +1,8 @@
 import 'dart:convert';
 import 'package:flutter/material.dart';
-import 'package:blurting/signup_questions/utils.dart';
+import 'package:blurting/pages/signup_questions/utils.dart';
 import 'package:blurting/token.dart';
-import 'package:blurting/signup_questions/sexual_preference.dart';
+import 'package:blurting/pages/signup_questions/sexual_preference.dart';
 import 'package:blurting/config/app_config.dart';
 import 'package:http/http.dart' as http;
 import 'package:blurting/Utils/provider.dart';
