+++ lib/pages/signup_questions/sexual_preference.dart
@@ -1,46 +1,35 @@
 import 'dart:convert';
 import 'package:blurting/Utils/provider.dart';
+import 'package:blurting/signup_questions/alcohol.dart';
 import 'package:flutter/material.dart';
 import 'package:blurting/token.dart';
 import 'package:http/http.dart' as http;
-import '../config/app_config.dart';
-import 'package:blurting/signup_questions/utils.dart';
-import 'package:blurting/signup_questions/mbti/mbti.dart';
 import 'package:blurting/utils/util_widget.dart';
-import 'major_list.dart';
+import 'package:blurting/pages/signup_questions/Utils.dart';
 
-class MajorPage extends StatefulWidget {
+import 'package:blurting/config/app_config.dart';
+
+class SexualPreferencePage extends StatefulWidget {
   final String selectedGender;
 
-  MajorPage({super.key, required this.selectedGender});
+  SexualPreferencePage({super.key, required this.selectedGender});
   @override
-  _MajorPageState createState() => _MajorPageState();
+  _SexualPreferencePageState createState() => _SexualPreferencePageState();
 }
 
-enum Major {
-  humanities, // 인문계열
-  social, // 사회계열
-  education, // 교육계열
-  engineering, // 공학계열
-  medical, // 의학계열
-  artsPhysical, // 예체능계열
-  naturalScience // 자연계열
-}
+enum SexualPreference { different, same, both, etc }
 
-class _MajorPageState extends State<MajorPage>
+class _SexualPreferencePageState extends State<SexualPreferencePage>
     with SingleTickerProviderStateMixin {
-  Major? _selectedMajor;
-  double _currentHeightValue = 160.0; // 초기 키 값
+  SexualPreference? _selectedSexPreference;
   AnimationController? _animationController;
   Animation<double>? _progressAnimation;
-  String? selectedMajor = majors[0];
-
   Future<void> _increaseProgressAndNavigate() async {
     await _animationController!.forward();
     Navigator.of(context).push(
       PageRouteBuilder(
         pageBuilder: (context, animation, secondaryAnimation) =>
-            MBTIPage(selectedGender: widget.selectedGender),
+            AlcoholPage(selectedGender: widget.selectedGender),
         transitionsBuilder: (context, animation, secondaryAnimation, child) {
           return FadeTransition(opacity: animation, child: child);
         },
@@ -49,6 +38,7 @@ class _MajorPageState extends State<MajorPage>
   }
 
   bool IsValid = false;
+
   @override
   void IsSelected() {
     IsValid = true;
@@ -59,21 +49,60 @@ class _MajorPageState extends State<MajorPage>
     super.initState();
 
     _animationController = AnimationController(
-      duration: Duration(milliseconds: 600), // 애니메이션의 지속 시간 설정
+      duration: Duration(milliseconds: 400), // 애니메이션의 지속 시간
       vsync: this,
     );
 
     _progressAnimation = Tween<double>(
-      begin: 8 / 15, // 시작 너비 (30%)
-      end: 9 / 15, // 종료 너비 (40%)
-    ).animate(
-        CurvedAnimation(parent: _animationController!, curve: Curves.easeInOut))
-      ..addListener(() {
-        setState(() {});
-      });
+      begin: 4 / 15, // 시작 게이지 값
+      end: 5 / 15, // 종료 게이지 값
+    ).animate(_animationController!);
+
+    _animationController?.addListener(() {
+      setState(() {}); // 애니메이션 값이 변경될 때마다 화면을 다시 그립니다.
+    });
   }
 
-  Widget major_checkbox(double width, Major major, String major_name) {
+  Future<void> _sendPostRequest() async {
+    var url = Uri.parse(API.signup);
+
+    var sexOrient = "";
+    if (_selectedSexPreference == SexualPreference.different) {
+      sexOrient = "hetero";
+    } else if (_selectedSexPreference == SexualPreference.same) {
+      sexOrient = "homo";
+    } else {
+      sexOrient = "bi";
+    }
+    String savedToken = await getToken();
+
+    var response = await http.post(
+      url,
+      headers: <String, String>{
+        'Content-Type': 'application/json; charset=UTF-8',
+        'Authorization': 'Bearer $savedToken',
+      },
+      body: json.encode({"sexOrient": sexOrient}), // JSON 형태로 인코딩
+    );
+
+    if (response.statusCode == 200 || response.statusCode == 201) {
+      // 서버로부터 응답이 성공적으로 돌아온 경우 처리
+
+      var data = json.decode(response.body);
+
+      if (data['signupToken'] != null) {
+        var token = data['signupToken'];
+
+        await saveToken(token);
+        _increaseProgressAndNavigate();
+      }
+    } else {
+      // 오류가 발생한 경우 처리
+    }
+  }
+
+  Widget sexualPreference_checkbox(double width,
+      SexualPreference sexualPreference, String sexualPreference_name) {
     return Container(
       width: width * 0.42, // 원하는 너비 값
       height: 48, // 원하는 높이 값
@@ -81,7 +110,7 @@ class _MajorPageState extends State<MajorPage>
         mainAxisSize: MainAxisSize.min,
         children: <Widget>[
           Checkbox(
-            value: _selectedMajor == major,
+            value: _selectedSexPreference == sexualPreference,
             side: BorderSide(color: Colors.transparent),
             fillColor: MaterialStateProperty.resolveWith<Color?>(
               (Set<MaterialState> states) {
@@ -93,7 +122,7 @@ class _MajorPageState extends State<MajorPage>
             ),
             onChanged: (bool? newValue) {
               setState(() {
-                _selectedMajor = major;
+                _selectedSexPreference = sexualPreference;
                 IsSelected();
               });
             },
@@ -102,12 +131,12 @@ class _MajorPageState extends State<MajorPage>
           GestureDetector(
             onTap: () {
               setState(() {
-                _selectedMajor = major;
+                _selectedSexPreference = sexualPreference;
                 IsSelected();
               });
             },
             child: Text(
-              major_name,
+              sexualPreference_name,
               style: TextStyle(
                 color: mainColor.black,
                 fontFamily: 'Pretendard',
@@ -121,57 +150,6 @@ class _MajorPageState extends State<MajorPage>
     );
   }
 
-  Future<void> _sendPostRequest() async {
-    print('_sendPostRequest called');
-    var url = Uri.parse(API.signup);
-    print(_selectedMajor);
-    var major = '';
-    if (_selectedMajor == Major.humanities) {
-      major = '인문계열';
-    } else if (_selectedMajor == Major.social) {
-      major = '사회계열';
-    } else if (_selectedMajor == Major.education) {
-      major = '교육계열';
-    } else if (_selectedMajor == Major.engineering) {
-      major = '공학계열';
-    } else if (_selectedMajor == Major.naturalScience) {
-      major = '자연계열';
-    } else if (_selectedMajor == Major.medical) {
-      major = '의학계열';
-    } else {
-      major = '예체능계열';
-    }
-
-    String savedToken = await getToken();
-    print(savedToken);
-
-    var response = await http.post(
-      url,
-      headers: <String, String>{
-        'Content-Type': 'application/json; charset=UTF-8',
-        'Authorization': 'Bearer $savedToken',
-      },
-      body: json.encode({"major": major}), // JSON 형태로 인코딩
-    );
-    print(response.body);
-    if (response.statusCode == 200 || response.statusCode == 201) {
-      // 서버로부터 응답이 성공적으로 돌아온 경우 처리
-      print('Server returned OK');
-      print('Response body: ${response.body}');
-      var data = json.decode(response.body);
-
-      if (data['signupToken'] != null) {
-        var token = data['signupToken'];
-        print(token);
-        await saveToken(token);
-        _increaseProgressAndNavigate();
-      } else {}
-    } else {
-      // 오류가 발생한 경우 처리
-      print('Request failed with status: ${response.statusCode}.');
-    }
-  }
-
   @override
   Widget build(BuildContext context) {
     Gender? _gender;
@@ -181,13 +159,13 @@ class _MajorPageState extends State<MajorPage>
       _gender = Gender.female;
     }
     double width = MediaQuery.of(context).size.width;
+
     return PopScope(
       canPop: true,
       onPopInvoked: (didPop) {
         sendBackRequest(context, false);
       },
       child: Scaffold(
-        resizeToAvoidBottomInset: false,
         backgroundColor: Colors.white,
         appBar: AppBar(
           backgroundColor: Colors.white,
@@ -207,38 +185,63 @@ class _MajorPageState extends State<MajorPage>
               SizedBox(
                 height: 50,
               ),
-              smallTitleQuestion("당신의 전공은 무엇인가요?"),
-              SizedBox(height: 30),
+              TitleQuestion("성적지향성은 어떻게 되시나요?"),
+              SizedBox(height: 30, width: 20),
               Row(
-                mainAxisAlignment: MainAxisAlignment.center, // 가로축 중앙 정렬
-                children: [
-                  major_checkbox(width, Major.humanities, "인문계열"),
-                  major_checkbox(width, Major.social, "사회계열"),
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: <Widget>[
+                  sexualPreference_checkbox(
+                      width, SexualPreference.different, "이성애자"),
+                  SizedBox(width: 23),
+                  sexualPreference_checkbox(
+                      width, SexualPreference.same, "동성애자"),
                 ],
               ),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center, // 가로축 중앙 정렬
-                children: [
-                  major_checkbox(width, Major.education, "교육계열"),
-                  major_checkbox(width, Major.engineering, "공학계열"),
-                ],
+              SizedBox(
+                height: 17,
               ),
               Row(
-                mainAxisAlignment: MainAxisAlignment.center, // 가로축 중앙 정렬
-                children: [
-                  major_checkbox(width, Major.naturalScience, "자연계열"),
-                  major_checkbox(width, Major.medical, "의학계열"),
-                ],
-              ),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center, // 가로축 중앙 정렬
-                children: [
-                  major_checkbox(width, Major.artsPhysical, "예체능계열"),
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: <Widget>[
+                  sexualPreference_checkbox(
+                      width, SexualPreference.both, "양성애자"),
+                  SizedBox(width: 23),
                   Container(
-                    width: width * 0.42, // 원하는 너비 값
-                    height: 48, // 원하는 높이 값
+                    width: width * 0.42,
+                    height: 48,
                   ),
                 ],
+              ),
+              SizedBox(height: 25),
+              Container(
+                width: 180,
+                height: 12,
+                child: RichText(
+                  text: TextSpan(
+                    style: TextStyle(
+                      fontSize: 10,
+                      fontWeight: FontWeight.w500,
+                      fontFamily: 'Pretendard',
+                      color: mainColor.black,
+                    ),
+                    children: [
+                      TextSpan(
+                        text: '*동성애자 선택시 ',
+                      ),
+                      TextSpan(
+                        text: '동성끼리만',
+                        style: TextStyle(
+                            color: Color(0xFFF66464)), // 원하는 색으로 변경하세요.
+                      ),
+                      TextSpan(
+                        text: ' 매칭해드립니다.',
+                      ),
+                    ],
+                  ),
+                ),
+              ),
+              SizedBox(
+                height: 21,
               ),
             ],
           ),
