+++ lib/pages/signup_questions/smoke.dart
@@ -2,11 +2,11 @@ import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 import 'package:blurting/token.dart';
-import 'package:blurting/signup_questions/Utils.dart';
-import 'package:blurting/signup_questions/height.dart';
+import 'package:blurting/pages/signup_questions/Utils.dart';
+import 'package:blurting/pages/signup_questions/height.dart';
 import 'package:blurting/Utils/provider.dart';
 import 'package:blurting/utils/util_widget.dart';
-import '../config/app_config.dart';
+import 'package:blurting/config/app_config.dart';
 
 final labels = ['안 피움', '가끔', '자주', '매일'];
 
@@ -128,7 +128,8 @@ class _SmokePageState extends State<SmokePage>
               SizedBox(
                 height: 25,
               ),
-              Center(child: ProgressBar(context, _progressAnimation!, _gender!)),
+              Center(
+                  child: ProgressBar(context, _progressAnimation!, _gender!)),
               SizedBox(
                 height: 50,
               ),
