+++ lib/pages/startpage/start_page.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:ui';
 import 'package:blurting/Utils/provider.dart';
-import 'package:blurting/signup_questions/phone_number.dart';
+import 'package:blurting/pages/signup_questions/phone_number.dart';
 import 'package:blurting/token.dart';
 import 'package:flutter/material.dart';
 import 'package:blurting/config/app_config.dart';
