+++ lib/token.dart
@@ -1,5 +1,5 @@
 import 'dart:convert';
-import 'package:blurting/startpage/start_page.dart';
+import 'package:blurting/pages/startpage/start_page.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
